+++ app/Services/OhDearDescriber.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Services;
+
+use App\Concerns\RendersBanner;
+use Illuminate\Console\Application;
+use NunoMaduro\LaravelConsoleSummary\Contracts\DescriberContract;
+use NunoMaduro\LaravelConsoleSummary\Describer;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class OhDearDescriber extends Describer
+{
+    use RendersBanner;
+
+    protected function describeTitle(Application $application, OutputInterface $output): DescriberContract
+    {
+        $this->renderBanner($output);
+
+        return parent::describeTitle($application, $output);
+    }
+}