+++ config/cache.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'default' => 'file',
+    'stores' => [
+        'file' => [
+            'driver' => 'file',
+            'path' => sys_get_temp_dir().'/ohdear',
+        ],
+    ],
+];