+++ config/commands.php
@@ -40,7 +40,7 @@ return [
     */
 
     'add' => [
-        // ..
+        //
     ],
 
     /*
@@ -61,6 +61,7 @@ return [
         Illuminate\Console\Scheduling\ScheduleRunCommand::class,
         Illuminate\Console\Scheduling\ScheduleListCommand::class,
         Illuminate\Console\Scheduling\ScheduleFinishCommand::class,
+        Illuminate\Foundation\Console\VendorPublishCommand::class,
         LaravelZero\Framework\Commands\StubPublishCommand::class,
     ],
 
@@ -76,7 +77,7 @@ return [
     */
 
     'remove' => [
-        // ..
+        //
     ],
 
 ];