+++ config/view.php
@@ -1,9 +1,0 @@
-<?php
-
-return [
-    'paths' => [
-        resource_path('views'),
-    ],
-
-    'compiled' => sys_get_temp_dir(),
-];