+++ resources/views/components/layouts/app.blade.php
@@ -1,3 +1,0 @@
-<div {{ $attributes->merge(['class' => 'ml-2 my-1']) }}>
-    {{ $slot }}
-</div>