+++ resources/views/notice.blade.php
@@ -1,3 +1,0 @@
-<x-layouts.app>
-    <span>{{ $notice }}</span>
-</x-layouts.app>