+++ resources/views/unauthorised.blade.php
@@ -1,9 +1,0 @@
-<x-layouts.app>
-    <div class="text-red-500">
-        <span>Unauthorised</span>
-    </div>
-
-    <p class="mt-1">
-        Please set the <span class="font-bold">OHDEAR_API_TOKEN</span> environment variable to your <a href="https://ohdear.app/user/api-tokens">Oh Dear API token</a>.
-    </p>
-</x-layouts.app>