+++ tests/TestCase.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Tests;
+
+use LaravelZero\Framework\Testing\TestCase as BaseTestCase;
+
+abstract class TestCase extends BaseTestCase {}